refactor(backend): extract gmail helpers (535 -> 385 lines)
